Exception handling

+++ src/main/java/com/dscomerce/dscomerce/services/ProductService.java
@@ -3,10 +3,14 @@ package com.dscomerce.dscomerce.services;
 import com.dscomerce.dscomerce.dto.ProductDTO;
 import com.dscomerce.dscomerce.entities.Product;
 import com.dscomerce.dscomerce.repository.ProductRepository;
+import com.dscomerce.dscomerce.services.exceptions.DatabaseException;
+import com.dscomerce.dscomerce.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -20,7 +24,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        return new ProductDTO( productRepository.findById(id).get());
+        return new ProductDTO( productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso nao encontrado")));
     }
 
 
@@ -45,9 +50,18 @@ public class ProductService {
         return new ProductDTO(entity);
     }
 
-    @Transactional
+
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product entity) {
